Extract marker rendering to a new module
It'll be easier to enable/disable it
